(PC-8354) Move BookingDetailsTicketContentProps into the components folder
